refactor: 필드에서 KakaoCredentials 사용하도록 변경
