+++ backend/src/main/java/zipgo/ZipgoApplication.java
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import zipgo.common.config.JwtCredentials;
-import zipgo.common.config.KakaoCredentials;
+import zipgo.auth.infra.kakao.config.KakaoCredentials;
 
 @EnableJpaAuditing
 @SpringBootApplication
