+++ backend/src/main/java/zipgo/auth/infra/kakao/config/KakaoCredentials.java
@@ -1,4 +1,4 @@
-package zipgo.common.config;
+package zipgo.auth.infra.kakao.config;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
